Updated infomation

+++ app/src/main/java/com/scootin/network/response/order/OrderHistoryItem.kt
@@ -2,6 +2,7 @@ package com.scootin.network.response.order
 
 import android.os.Parcelable
 import com.scootin.network.response.AddressDetails
+import com.scootin.network.response.UserInfo
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
@@ -9,7 +10,7 @@ data class OrderHistoryItem(
     val addressDetails: AddressDetails,
     val directOrder: Boolean,
     val expressDelivery: Boolean,
-    val id: Int,
+    val id: Long,
     val orderDate: String,
     val orderStatus: String,
     val rejectReason: String,
